admin authentication added with forgot password functionality

+++ resources/views/partials/posts.blade.php
@@ -38,8 +38,10 @@
             @if (count($post->comments) > 1)
             <a id="toggleCommentBtn{{ $post->id }}" href="" onclick="toggleComments(event, {{ $post->id }});">view previous comments</a>
             @endif
-            <a style="margin:0px 10px;" href="" onclick="showEditReviewModal(event, {{$post->id}})">Edit</a>
-            <a href="#" onclick="deletePost(event, {{ $post->id }})">Delete</a>
+            @if (Auth::user()->id == $post->user->id)
+              <a style="margin:0px 10px;" href="" onclick="showEditReviewModal(event, {{$post->id}})">Edit</a>
+              <a href="#" onclick="deletePost(event, {{ $post->id }})">Delete</a>
+            @endif
           </span>
         </p>
